+++ Mindkeepr/models/__init__.py
@@ -1,0 +1,3 @@
+from .category import *
+from .location import *
+from .models import *
